feat: add tests for the services

+++ src/application/useCases/Location/DeleteLocationService.ts
@@ -3,6 +3,7 @@ import 'reflect-metadata'
 import { IDeleteLocationService } from "../../../domain/useCases/Location/DeleteLocationService";
 import ILocationRepository from "../../../domain/repositories/Location";
 import { LocationServiceDTO } from "../../../domain/useCases/Location/Location";
+import ApplicationError from "../../../domain/errors/ApplicationError";
 
 @injectable()
 export default class DeleteLocationsService implements IDeleteLocationService {
@@ -11,6 +12,8 @@ export default class DeleteLocationsService implements IDeleteLocationService {
     ){}
 
     async execute(input: LocationServiceDTO.DeleteLocationInput): Promise<void> {
+        const exists = await this.locationRepository.getById(input.locationId)
+        if (!exists) throw new ApplicationError("This location doesn't exists", 400)
         await this.locationRepository.delete(input.locationId)
     }
 }
